+++ database/migrations/2026_03_11_101629_add_foreign_keys_to_intern_tasks_table.php
@@ -11,10 +11,12 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::table('intern_tasks', function (Blueprint $table) {
-            $table->foreign(['eti_id'], 'intenrkey')->references(['eti_id'])->on('intern_accounts')->onUpdate('restrict')->onDelete('restrict');
-            $table->foreign(['assigned_by'], 'tasksupkey')->references(['manager_id'])->on('manager_accounts')->onUpdate('restrict')->onDelete('restrict');
-        });
+        if (Schema::hasTable('intern_tasks')) {
+            Schema::table('intern_tasks', function (Blueprint $table) {
+                $table->foreign(['eti_id'], 'intenrkey')->references(['eti_id'])->on('intern_accounts')->onUpdate('restrict')->onDelete('restrict');
+                $table->foreign(['assigned_by'], 'tasksupkey')->references(['manager_id'])->on('manager_accounts')->onUpdate('restrict')->onDelete('restrict');
+            });
+        }
     }
 
     /**
