+++ database/migrations/2026_03_11_101629_add_foreign_keys_to_manager_roles_table.php
@@ -11,9 +11,11 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::table('manager_roles', function (Blueprint $table) {
-            $table->foreign(['manager_id'])->references(['manager_id'])->on('manager_accounts')->onUpdate('cascade')->onDelete('cascade');
-        });
+        if (Schema::hasTable('manager_roles')) {
+            Schema::table('manager_roles', function (Blueprint $table) {
+                $table->foreign(['manager_id'])->references(['manager_id'])->on('manager_accounts')->onUpdate('cascade')->onDelete('cascade');
+            });
+        }
     }
 
     /**
